Add configuration `matlab_short_links` (#171)
* Added new configuration `matlab_short_links`

Working on getting more MATLAB like rendering and links.
Rather than `target.+package.ClassBar`, it will displayed and referenced
as `package.ClassBar`.

Major rewrite of parsing for MATLAB files in `matlab_src_dir`. It's now done
at startup and stored in a dict.

Added a lot of tests to verify the new functionality.