Add option `matlab_auto_link` (#183)
* Add linking of classes and functions in See also lines.

* Add option `matlab_auto_link`

Currently just True or False, default is False.

In the future we may want to change this to a string with possible values like "see_also", "all" and maybe something inbetween,

* Add ref_role() method to MatObject

Returns a string with the role to be used for references to this type of object (e.g. "class", "func", "meth", "attr").

* Add option to auto-link everywhere.

Changed `matlab_auto_link` option to string value, currently with "see_also" or "all" as valid options.

* Exclude auto-linking class name in "<class> Properties:" or "<class> Methods:".

These are headings in class docstrings that are recognized by MATLAB's `help` and `doc`.

* Fix issue with matching things that are not whole words.

* Allow case-insensitive match of "See Also" w or w/o trailing ":"

* Add tests for matlab_auto_link = "see_also"

* Add code to force matlab_short_links to True if matlab_auto_link is not None.

* Fix issue with broken auto-links if matlab_short_links is False

Add negative look-ahead for ` or . or +

Avoids turning ...
:class:`target.myClass`
... into...
:class:`target.:class:`myClass``
etc.

* Revert "Add code to force matlab_short_links to True if matlab_auto_link is not None."

This reverts commit 91746429a84548289b5eec41a9713b4511c1a1d4.

* Refactor "see also" auto-linking.

Look for each entity on “see also” line in the table, rather than the other way around.
Adds double-backquoting of unknown entities.

* Update tests to include MATLAB names in See also lines.

* Add test of double-backquotes around unknown entities in see also sections.

* Create entities_name_map only once at the same time as entities_table.

* Modify logic so that matlab_auto_link = "all" handles see_also line first.

... including wrapping unknown names in double-backquotes.

* Refactor MatlabDocumenter.auto_link() into separate methods

- auto_link_see_also()
- auto_link_all()

* Refactor make_baseclass_links() into 2 new MatClass methods

- fullname()
- link()

* Change "see_also" value of matlab_auto_link to "basic".

* Add auto-linking of property and method names in class docstring.

Only affects names in lines following "<class> Properties:" or "<class> Methods:"

* Remove leading . in MatClass.fullname() results

* Fix issue with matlab_auto_link = "all" trying to link class names within existing links.

* Update tests for matlab_auto_link = "basic" to include property/method names in class docstrings.

* Add a test for matlab_auto_link = "all"

* Update README.rst for latest auto-link behavior.

* Add parens following the name of a linked method.

* Add auto-linking of method names in class/method/property docstrings.

Requires matlab_auto_link = "all" and the method name must appear with a trailing () to be auto-linked.

* Add autodoc tests for matlab_short_links = True.

* Fix bug in class links when matlab_short_links = True

This affected both base class links and auto-linking. The issue was that package names were not being included in the link target.

* Fix class links so names are consistent with targets.

This makes the visible names consistent with the `matlab_short_links` and `matlab_keep_package_prefix` options.

* Skip over literal blocks when auto-linking with matlab_auto_link = "all"