Minor tweaks for cleanliness' sake (#283)
* chore: Remove unused imports

* chore: Remove pass-through imports

`MatModule` and `MatObject` were imported, but never used, in the
`mat_documenters` module; however, they were passed through to the
`matlab` and `test_matlabify` modules.  This is unnecessary and
confusing for readers, so this commit removes the imports from
`mat_documenters`, and instead imports them directly from `mat_types`
where needed.

* chore: Remove unused variables

* fix: Remove accidental =

Having `docstring == None` here did effectively nothing (it returned
`False`, but nothing was done with that return value).  Given the
comment above, it looks like this was just a typo, which this commit
fixes.

* refactor: Rename _ -> translation

`sphinx.locale` provides `_` as an alias to `get_translation("sphinx")`
for brevity's sake, but using `_` is problematic, because (1) its
meaning is opaque to the reader, and (2) it's a common Python practice
to use `_` for "throw-away" variables (those that are returned by a
function or generator, but are unneeded and won't be used), and doing so
will shadow the import (as was the case here in the `clear_doc` method).

---------

Co-authored-by: Remi Gau <remi_gau@hotmail.com>