fix: Handle the zero matches case
When trying to test out the `master` branch against my project, to see
if we're ready for a 0.23.0 release, I ran into the following error:

    File ".../sphinx/events.py", line 415, in emit
        raise ExtensionError(
        ...<4 lines>...
        ) from exc
    sphinx.errors.ExtensionError: Handler <function analyze at 0x...>
        for event 'builder-inited' threw an exception (exception: list
        index out of range)

Digging further, the problem was actually:

    File ".../mat_tree_sitter_parser.py", line 373, in
        _parse_argument_section
        _, arg_match = q_arg.matches(arg)[0]
                       ~~~~~~~~~~~~~~~~~~^^^
    IndexError: list index out of range

I was able to rewind through history to find this problem first arose
with commit 4cd5c5136b92bdfda6cf6a182bbcdcf12a23058d.  The problem is
occasionally `q_arg.matches(arg)` will return no matches (an empty
list), so we need to check for that possibility before attempting to
access the first element of the list.